// autogpt_platform/frontend/src/app/api/__generated__/models/agentSavedResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { ResponseType } from "./responseType";
import type { AgentSavedResponseSessionId } from "./agentSavedResponseSessionId";

/**
 * Response when an agent is saved to the library.
 */
export interface AgentSavedResponse {
  type?: ResponseType;
  message: string;
  session_id?: AgentSavedResponseSessionId;
  agent_id: string;
  agent_name: string;
  library_agent_id: string;
  library_agent_link: string;
  agent_page_link: string;
}