// autogpt_platform/frontend/src/app/api/__generated__/models/clarificationNeededResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { ResponseType } from "./responseType";
import type { ClarificationNeededResponseSessionId } from "./clarificationNeededResponseSessionId";
import type { ClarifyingQuestion } from "./clarifyingQuestion";

/**
 * Response when the LLM needs more information from the user.
 */
export interface ClarificationNeededResponse {
  type?: ResponseType;
  message: string;
  session_id?: ClarificationNeededResponseSessionId;
  questions?: ClarifyingQuestion[];
}