// autogpt_platform/frontend/src/app/api/__generated__/models/errorResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { ResponseType } from "./responseType";
import type { ErrorResponseSessionId } from "./errorResponseSessionId";
import type { ErrorResponseError } from "./errorResponseError";
import type { ErrorResponseDetails } from "./errorResponseDetails";

/**
 * Response for errors.
 */
export interface ErrorResponse {
  type?: ResponseType;
  message: string;
  session_id?: ErrorResponseSessionId;
  error?: ErrorResponseError;
  details?: ErrorResponseDetails;
}